+++ src/ht_manager/bot/client.py
@@ -11,15 +11,19 @@ from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 from ht_manager.bot.commands.addctf import register_addctf_command
 from ht_manager.bot.commands.addctfmember import register_addctfmember_command
 from ht_manager.bot.commands.addresult import register_addresult_command
+from ht_manager.bot.commands.archivectf import register_archivectf_command
 from ht_manager.bot.commands.ctfmembers import register_ctfmembers_command
+from ht_manager.bot.commands.deletectf import register_deletectf_command
 from ht_manager.bot.commands.editctf import register_editctf_command
 from ht_manager.bot.commands.editresult import register_editresult_command
+from ht_manager.bot.commands.endctf import register_endctf_command
 from ht_manager.bot.commands.nextctf import register_nextctf_command
 from ht_manager.bot.commands.participation import register_participation_command
 from ht_manager.bot.commands.ping import register_ping_command
 from ht_manager.bot.commands.removectfmember import register_removectfmember_command
 from ht_manager.bot.commands.resolvepoll import register_resolvepoll_command
 from ht_manager.bot.commands.resultsync import register_resultsync_command
+from ht_manager.bot.commands.setcategory import register_setcategory_command
 from ht_manager.bot.commands.setupctf import register_setupctf_command
 from ht_manager.config import Settings
 from ht_manager.jobs.cleanup import cleanup_expired_resources
@@ -59,6 +63,10 @@ class HTManagerBot(commands.Bot):
         register_addresult_command(self)
         register_editresult_command(self)
         register_resultsync_command(self)
+        register_deletectf_command(self)
+        register_setcategory_command(self)
+        register_endctf_command(self)
+        register_archivectf_command(self)
         self.tree.on_error = self._on_app_command_error
         guild = discord.Object(id=self.settings.discord_guild_id)
         self.tree.copy_global_to(guild=guild)